Use atomic for IsInitialBlockDownload()
use an atomic bool for fIsInitialBlockDownload and then reset the flag
every time a block is processed or a header is received.
